# src/main/java/com/accountapi/dto/auth/AuthResponseDto.java
package com.accountapi.dto.auth;

import lombok.Getter;

@Getter
public class AuthResponseDto {
    private String token;

    public AuthResponseDto(String token) {
        this.token = token;
    }
}