# src/main/java/com/accountapi/dto/transfer/TransferRequestDto.java
package com.accountapi.dto.transfer;

import jakarta.validation.constraints.DecimalMin;
import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;
import java.math.BigDecimal;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class TransferRequestDto {

    @NotNull
    private Long to;

    @DecimalMin(value = "0.01", inclusive = true)
    private BigDecimal value;
}